Create new component with doughnut graph
