+++ src/components/DisplayPerformanceData.jsx
@@ -1,7 +1,6 @@
 import React, { Component } from "react";
 import { getData } from "../modules/performanceData";
-import { Doughnut } from "react-chartjs-2";
-
+import DisplayDoughnutChart from "../components/DisplayDoughnutChart";
 class DisplayPerformanceData extends Component {
   state = {
     performanceData: null,
@@ -25,37 +24,7 @@ class DisplayPerformanceData extends Component {
   }
 
   render() {
-    const doughnutData = {};
-
-    if (this.state.performanceData != null) {
-      this.state.performanceData.forEach((entry) => {
-        let label = entry.data.message;
-        if (doughnutData[label] == null) {
-          doughnutData[label] = 0;
-        }
-        doughnutData[label] += 1;
-      });
-    }
-
-    const data = {
-      labels: Object.keys(doughnutData),
-      datasets: [
-        {
-          data: Object.values(doughnutData),
-          backgroundColor: [
-            "rgba(205, 92, 92)",
-            "rgba(240, 128, 128)",
-            "rgba(250, 128, 114)",
-            "rgba(233, 150, 122)",
-            "rgba(255, 160, 122)",
-          ],
-        },
-      ],
-    };
-
-    let doughnut = <Doughnut data={data} />;
-
-    return <div id="index"><h1>Accumulated results</h1>{doughnut}</div>;
+    return <DisplayDoughnutChart performanceData={this.state.performanceData}/>
   }
 }
